requirements.txt
# SmartEye Backend API Requirements

# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data validation and models
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML libraries
torch>=2.0.0
torchvision>=0.15.0
doclayout-yolo
transformers>=4.35.0
huggingface-hub>=0.19.0

# OpenAI API
openai>=1.3.0

# OCR
pytesseract>=0.3.10
tesseract

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# PDF processing  
PyMuPDF>=1.23.0

# File processing
python-magic>=0.4.27

# System monitoring
psutil>=5.9.0

# Async and concurrency
asyncio
aiofiles>=23.0.0

# Logging and monitoring
loguru>=0.7.0

# Database (optional)
redis>=5.0.0
asyncpg>=0.29.0  # for PostgreSQL
sqlalchemy>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # for testing

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
rich>=13.6.0